Refactor components/App
