+++ components/App/App.js
@@ -7,17 +7,17 @@ import Head from 'next/head';
 import { Provider } from 'react-redux';
 import Router from 'next/router';
 import { fromJS } from 'immutable';
-import { setLogin } from '../util/gql';
-import configure from '../redux';
-import { showDialog, load } from '../redux/auth';
+import { setLogin } from '../../util/gql';
+import configure from '../../redux';
+import { showDialog, load } from '../../redux/auth';
 import AppHeader from './AppHeader';
 import AppFooter from './AppFooter';
-import LoginModal from './Modal/LoginModal';
+import LoginModal from '../Modal/LoginModal';
 import moment from 'moment';
 import 'moment/locale/zh-tw';
 import style from './App.css';
 import NProgress from 'nprogress';
-const { GA_TRACKER, AUTOTRACK_FILENAME } = require('../config');
+const { GA_TRACKER, AUTOTRACK_FILENAME } = require('../../config');
 
 let isBootstrapping = true;
 moment.locale('zh-tw');
