+++ components/App/AppHeader.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { EDITOR_FACEBOOK_GROUP } from '../constants/urls';
-import { Link } from '../routes';
-import { showDialog, logout } from '../redux/auth';
+import { EDITOR_FACEBOOK_GROUP } from '../../constants/urls';
+import { Link } from '../../routes';
+import { showDialog, logout } from '../../redux/auth';
 
 function AppHeader({ user, onLoginClick, onLogoutClick }) {
   return (
