+++ components/App/index.js
@@ -1,0 +1,2 @@
+import app from './App';
+export default app;
